fix(dashboard/courses): :bug: fix optimisticCourses & fix loading the wrong courses in published and drafts tabs

+++ providers/CourseState.provider.tsx
@@ -21,11 +21,14 @@ import { toast } from "sonner";
 import { getBundles as fetchBundles } from "@/actions/bundles.actions";
 import { useFormState } from "react-dom";
 import { CoursesFilter, Tab } from "@/types/manage.courses.types";
+import { useSearchParams } from "next/navigation";
+
+type isSelected = Record<number, boolean> | undefined;
 
 type State = {
   activeTab: Tab;
-  isCourseSelected: Record<number, boolean> | undefined;
-  isBundleSelected: Record<number, boolean> | undefined;
+  isCourseSelected: isSelected;
+  isBundleSelected: isSelected;
   courseInfo: CourseWithSafeFellow | undefined;
   filter: CoursesFilter;
   fellow: SafeUser | undefined;
@@ -36,21 +39,16 @@ type State = {
 
 type Action =
   | { type: "SET_ACTIVE_TAB"; payload: Tab }
-  | {
-      type: "SET_COURSE_SELECTED";
-      payload: Record<number, boolean> | undefined;
-    }
-  | {
-      type: "SET_BUNDLE_SELECTED";
-      payload: Record<number, boolean> | undefined;
-    }
-  | { type: "SET_COURSE_INFO"; payload: CourseWithSafeFellow | undefined }
   | { type: "SET_FILTER"; payload: CoursesFilter }
+  | { type: "SET_COURSE_SELECTED"; payload: isSelected }
+  | { type: "SET_BUNDLE_SELECTED"; payload: isSelected }
+  | { type: "SET_COURSE_INFO"; payload: CourseWithSafeFellow | undefined }
   | { type: "SET_FELLOW"; payload: SafeUser | undefined }
   | { type: "SET_LOADING"; payload: boolean }
   | { type: "SET_COURSES"; payload: CourseWithSafeFellow[] }
   | { type: "SET_BUNDLES"; payload: BundleWithCoursesNames[] }
-  | { type: "DELETE_COURSE"; payload: number };
+  | { type: "DELETE_COURSE"; payload: number }
+  | { type: "CLEAR_COURSES" };
 
 const initialState: State = {
   activeTab: "published",
@@ -81,16 +79,17 @@ function reducer(state: State, action: Action): State {
     case "SET_LOADING":
       return { ...state, isLoading: action.payload };
     case "SET_COURSES":
-      return { ...state, courses: action.payload };
+      return { ...state, courses: action.payload, isLoading: false };
     case "SET_BUNDLES":
-      return { ...state, bundles: action.payload };
+      return { ...state, bundles: action.payload, isLoading: false };
     case "DELETE_COURSE":
       return {
         ...state,
-        courses: state.courses?.filter(
-          (course) => course.id !== action.payload,
-        ),
+        courses: state.courses.filter((course) => course.id !== action.payload),
       };
+    case "CLEAR_COURSES":
+      return { ...state, courses: [] };
+
     default:
       return state;
   }
@@ -113,31 +112,27 @@ export const CourseStateProvider = ({ children }: { children: ReactNode }) => {
   const [deleteState, deleteAction] = useFormState(deleteCourse, {});
   const [isPending, startTransition] = useTransition();
 
-  const [optimisticCourses, addOptimisticCourse] = useOptimistic(
-    state.courses,
-    (currentState, courseIdToRemove: number | null) => {
-      return currentState.filter((course) => course.id !== courseIdToRemove);
-    },
-  );
+  const getCourses = useCallback(
+    async (filter: CoursesFilter) => {
+      if (!state.filter) return;
 
-  const getCourses = useCallback(async () => {
-    if (!state.filter) return;
+      try {
+        dispatch({ type: "SET_LOADING", payload: true });
+        dispatch({ type: "CLEAR_COURSES" });
 
-    try {
-      dispatch({ type: "SET_LOADING", payload: true });
-      const data = await getSafeCourses(state.filter);
-      if (data.error) throw new Error(data.message);
-      if (data.courses) {
-        dispatch({ type: "SET_COURSES", payload: data.courses });
+        const data = await getSafeCourses(filter);
+        if (data.error) throw new Error(data.message);
+        if (data.courses) {
+          dispatch({ type: "SET_COURSES", payload: data.courses });
+        }
+      } catch (error) {
+        toast.error(
+          error instanceof Error ? error.message : "Failed to fetch courses",
+        );
       }
-    } catch (error) {
-      toast.error(
-        error instanceof Error ? error.message : "Failed to fetch courses",
-      );
-    } finally {
-      dispatch({ type: "SET_LOADING", payload: false });
-    }
-  }, [state.filter]);
+    },
+    [state.filter],
+  );
 
   const getBundles = useCallback(async () => {
     try {
@@ -150,37 +145,77 @@ export const CourseStateProvider = ({ children }: { children: ReactNode }) => {
       toast.error(
         error instanceof Error ? error.message : "Failed to fetch bundles",
       );
-    } finally {
-      dispatch({ type: "SET_LOADING", payload: false });
     }
   }, []);
 
   const forceUpdateCourses = useCallback(async () => {
-    await getCourses();
-  }, [getCourses]);
+    await getCourses(state.filter);
+  }, [getCourses, state.filter]);
 
   const forceUpdateBundles = useCallback(async () => {
     await getBundles();
   }, [getBundles]);
 
+  const searchParams = useSearchParams();
+  const activeTabParam = searchParams?.get("tab");
+
   useEffect(() => {
-    if (state.activeTab === "bundles") getBundles();
-  }, [getBundles, state.activeTab]);
+    if (state.activeTab === "bundles") {
+      getBundles();
+    } else if ((state.activeTab && activeTabParam) === "published") {
+      getCourses(state.filter);
+    } else if (state.activeTab && activeTabParam === "draft") {
+      getCourses("draft");
+    } else {
+      getCourses(state.filter);
+    }
+  }, [activeTabParam, getBundles, getCourses, state.activeTab, state.filter]);
+
+  const [optimisticCourses, addOptimisticCourse] = useOptimistic(
+    state.courses,
+    (
+      currentState,
+      update: {
+        type: "remove" | "restore";
+        courseId: number;
+      },
+    ) => {
+      switch (update.type) {
+        case "remove":
+          return state.courses.filter(
+            (course) => course.id !== update.courseId,
+          );
+        case "restore":
+          const courseToRestore = state.courses.find(
+            (course) => course.id === update.courseId,
+          );
+          return courseToRestore
+            ? [...currentState, courseToRestore]
+            : currentState;
+
+        default:
+          return currentState;
+      }
+    },
+  );
 
   const handleDelete = useCallback(
     (courseId: number) => {
+      const courseToDelete = state.courses.find(
+        (course) => course.id === courseId,
+      );
+      if (!courseToDelete) return toast.error("Course is not available!");
+
       startTransition(() => {
-        addOptimisticCourse(courseId);
+        addOptimisticCourse({ type: "remove", courseId });
         const formData = new FormData();
         formData.append("courseId", courseId.toString());
         deleteAction(formData);
       });
     },
-    [addOptimisticCourse, deleteAction],
+    [addOptimisticCourse, deleteAction, state.courses],
   );
-
   const toastId = useRef<string | number>();
-
   useEffect(() => {
     if (isPending) toastId.current = toast.loading("Loading...");
   }, [isPending]);
@@ -190,15 +225,19 @@ export const CourseStateProvider = ({ children }: { children: ReactNode }) => {
       toast.success(deleteState.message, { id: toastId.current });
       dispatch({
         type: "DELETE_COURSE",
-        payload: Number(deleteState?.deletedId),
+        payload: Number(deleteState.deletedId),
       });
     }
-    if (deleteState?.error) toast.error(deleteState.message);
-  }, [deleteState]);
-
-  useEffect(() => {
-    getCourses();
-  }, [getCourses, state.filter]);
+    if (deleteState?.error) {
+      toast.error(deleteState.message, { id: toastId.current });
+      startTransition(() => {
+        addOptimisticCourse({
+          type: "restore",
+          courseId: Number(deleteState?.deletedId),
+        });
+      });
+    }
+  }, [deleteState, addOptimisticCourse]);
 
   return (
     <CourseStateContext.Provider
